feat: Add SvVariable struct to SvModule
This commit adds the `SvVariable` struct to the `SvModule` in order to support variable declarations in SystemVerilog modules. The `SvVariable` struct represents a variable identifier and can be accessed through the `variables` field of the `SvModule` class.

This change enhances the functionality of the `SvModule` class by allowing it to store and retrieve variable declarations, providing more comprehensive representation of SystemVerilog modules.